Реализация сортировки по возрасту плюс тесты
